+++ code/cell.electron/app/src/main/common/libs.ts
@@ -1,3 +1,5 @@
+export { Subject, Observable } from 'rxjs';
+
 export { fs } from '@platform/fs';
 export { Schema, Urls, Uri } from '@platform/cell.schema';
 export { Client, HttpClient } from '@platform/cell.client';
